requirements.txt
gradio
whisper
groq
python-dotenv
gtts
